driver tests only use the new framework
