+++ crates/driver/src/infra/config/mod.rs
@@ -1,6 +1,9 @@
 use {
     self::file::ContractsConfig,
-    crate::infra::{liquidity, mempool, simulator, solver},
+    crate::{
+        domain::eth,
+        infra::{liquidity, mempool, simulator, solver},
+    },
 };
 
 pub mod file;
@@ -9,6 +12,7 @@ pub mod file;
 #[derive(Debug)]
 pub struct Config {
     pub disable_access_list_simulation: bool,
+    pub disable_gas_simulation: Option<eth::Gas>,
     pub solvers: Vec<solver::Config>,
     pub liquidity: liquidity::Config,
     pub tenderly: Option<simulator::tenderly::Config>,
