+++ crates/driver/src/infra/solver/dto/auction.rs
@@ -149,9 +149,6 @@ struct Order {
     fee_amount: eth::U256,
     kind: Kind,
     partially_fillable: bool,
-    // TODO There is no executed field here. Is this an oversight? For partial orders, should the
-    // driver reduce the buy/sell amount by the executed amounts before sending the order to the
-    // solver?
     class: Class,
 }
