+++ crates/driver/src/tests/cases/example_config.rs
@@ -1,4 +1,4 @@
-use crate::tests::setup;
+use crate::tests;
 
 /// Test that the example configuration file is valid by checking that the
 /// driver does not crash when started with this file.
@@ -6,5 +6,5 @@ use crate::tests::setup;
 #[ignore]
 async fn test() {
     let example_config_file = std::env::current_dir().unwrap().join("example.toml");
-    setup().config(example_config_file).done().await;
+    tests::setup().config(example_config_file).done().await;
 }
