+++ crates/driver/src/tests/mod.rs
@@ -2,7 +2,7 @@ mod boundary;
 mod cases;
 mod setup;
 
-pub use setup::new::setup;
+pub use setup::setup;
 
 fn hex_address(value: ethcontract::H160) -> String {
     format!("0x{}", hex::encode(value.as_bytes()))
